Scraper
Scraper changes and minor algorithm changes.
